combine upload data function into visualize_openfast
